Changements des layouts, ajouts de boites de confirmations, layouts marche sur light and night mode, fonctionnalité de deconnexion et enlevage du back button du téléphone android.

+++ app/src/main/java/com/example/griffinpiece/adapters/ShowsFavoritesRecyclerViewAdapter.kt
@@ -1,12 +1,16 @@
 package com.example.griffinpiece.adapters
 
 import android.annotation.SuppressLint
+import android.content.DialogInterface
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.appcompat.app.AlertDialog
 import androidx.core.os.bundleOf
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
@@ -14,6 +18,7 @@ import com.example.griffinpiece.R
 import com.example.griffinpiece.models.Show
 import com.example.griffinpiece.ui.favori.FavoritesViewModel
 import com.squareup.picasso.Picasso
+
 
 class ShowsFavoritesRecyclerViewAdapter(private val datasetShows: MutableList<Show>, val favoritesViewModel: FavoritesViewModel) :
     RecyclerView.Adapter<ShowsFavoritesRecyclerViewAdapter.ShowViewHolder>() {
@@ -38,15 +43,28 @@ class ShowsFavoritesRecyclerViewAdapter(private val datasetShows: MutableList<Sh
         }
         val btnFavorite = holderShow.view.findViewById<Button>(R.id.btnFvorite)
         btnFavorite.setOnClickListener {
-            val id = datasetShows[position].id
-            this.favoritesViewModel.id = id
-            println(this.favoritesViewModel.id)
-            this.favoritesViewModel.delete()
+            val builder = AlertDialog.Builder(holderShow.view.context)
+            builder.setMessage("Enlever l'émission des favoris?")
+                .setCancelable(false)
+                .setPositiveButton("Oui") { dialog, which ->
+                    val id = datasetShows[position].id
+                    this.favoritesViewModel.id = id
+                    println(this.favoritesViewModel.id)
+                    this.favoritesViewModel.delete()
 
-            datasetShows.removeAt(position)
-            this.notifyDataSetChanged()
+                    datasetShows.removeAt(position)
+                    this.notifyDataSetChanged()
+                }
+                .setNegativeButton("Non") { dialog, which ->
+                    dialog.dismiss()
+                }
+                val alert = builder.create()
+                alert.show()
+                val nbutton = alert.getButton(DialogInterface.BUTTON_NEGATIVE)
+                nbutton.setTextColor(Color.parseColor("#FF5722"))
+                val pbutton = alert.getButton(DialogInterface.BUTTON_POSITIVE)
+                pbutton.setTextColor(Color.parseColor("#FF5722"))
         }
-
     }
 
     override fun getItemCount(): Int = this.datasetShows.size
